Billing_service/billing/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
distlib==0.3.6
Django==4.2.1
django-split-settings==1.2.0
djangorestframework==3.14.0
drf-yasg==1.21.5
gunicorn==20.1.0
Faker==18.6.2
filelock==3.12.0
flake8==6.0.0
identify==2.5.24
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.1
platformdirs==3.5.1
pre-commit==3.3.1
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.5.5
requests==2.30.0
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0